+++ api/setup.go
@@ -17,6 +17,7 @@ func Setup(app fiber.Router) {
 
 	api.Post("/login", routes.Login)
 	api.Get("/logout", routes.Logout)
+	api.Post("/update-password", auth.Middleware(), routes.UpdatePassword)
 
 	api.Post("/search", auth.Middleware(), routes.Search)
 	api.Get("/stats", auth.Middleware(), routes.Stats)
